refactor : move np_profile_dialog to it's own folder
