+++ src/core/libraries/libs.cpp
@@ -38,7 +38,7 @@
 #include "core/libraries/np/np_matching2.h"
 #include "core/libraries/np/np_partner.h"
 #include "core/libraries/np/np_party.h"
-#include "core/libraries/np/np_profile_dialog.h"
+#include "core/libraries/np/np_profile_dialog/np_profile_dialog.h"
 #include "core/libraries/np/np_score.h"
 #include "core/libraries/np/np_sns_facebook_dialog.h"
 #include "core/libraries/np/np_trophy.h"
