+++ src/core/libraries/np/np_profile_dialog/np_profile_dialog.h
